Encapsulate functionality to supplementary_plot_helpers.py
